Create Oracle-RPD-Preparation.ps1
the issue that the script solves is when in multi Oracle BI environments we have the developer modifies the Oracle RPD with the new changes solving a bug or creating new functionality and so on. so this RPD file passed to the operation team to prepare this RPD file and then the operation team prepare it with tow major changes RPD environment variables and RPD connection pools.this manual effort takes a lot of effort and also the probability of human mistakes is very high. in my case there were around 15 RPD environment variables and more than 25 connection pool. so I've developed a solution which saves a lot of time if the manual effort takes around an hour my solution will take around 18 seconds.my solution is a powershell script which reads the environment variables and configurations of connection pools from excel sheet which resides on a sharepoint and then create a command file then run the command file against the RPD to get it prepared for the target environment.

the below are steps which script takes to do the preparation.
-ask the user about sharepoint username and password to download the excel sheet 
-convert the sheet to csv files according to the selected environment 
-generate the environment variables section in the RPD command file 
-generate the connection pools details according to the csv file generate 
-update the RPD file 
-exit the client tool